jim: garbage collection fix
Improve the check for an object that exists only in the command table with
reference count of one. The object being checked needs to be the same object
as the one in the command table. And also objects of type reference can
be in the command table so check those too.

Fixes #245

Signed-off-by: Steve Bennett <steveb@workware.net.au>